qa/cephfs: move tests related to CloneProgressReporter to a new file
So that there is a separate file where several different groups of tests
for CloneProgressReporter can reside together.
