+++ qa/suites/fs/volumes/tasks/volumes/test/clone-progress.yaml
@@ -2,4 +2,4 @@ tasks:
   - cephfs_test_runner:
       fail_on_skip: false
       modules:
-        - tasks.cephfs.test_volumes.TestCloneProgressReporter
+        - tasks.cephfs.volumes.test_clone_stats.TestCloneProgressReporter
